Separating packing algorithms
