+++ src/generator/packing/compact-packing-algorithm.js
@@ -1,0 +1,5 @@
+import Blueprint from 'factorio-blueprint';
+import createLocalConnections from './create-local-connections';
+
+export default ({clock, signals, timers, states}) => {
+};
