+++ src/generator/packing/packing-algorithms.js
@@ -1,0 +1,7 @@
+import compactPackingAlgorithm from './compact-packing-algorithm';
+import debugPackingAlgorithm from './debug-packing-algorithm';
+
+export default {
+    compactPackingAlgorithm,
+    debugPackingAlgorithm
+};
